singular-agents/co-borrower/requirements.txt
# Core Framework
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings

# LangChain
langchain
langchain-core
langchain-google-genai

# Google AI
google-generativeai
langchain-groq                    # ⚠️ CRITICAL - MISSING IN YOUR FILE

# PDF Processing
PyMuPDF
Pillow

# Utilities
python-dotenv
tenacity
werkzeug

# Async support
aiofiles

# HTTP requests
requests
httpx

# Monitoring & Logging (Production)
python-json-logger

# Testing (Optional - can remove in production)
pytest
pytest-asyncio

# Performance
orjson
